app/bitget/config/keys.py
import os
from dotenv import load_dotenv

load_dotenv()

BITGET_API_KEY = os.getenv("BITGET_API_KEY", "")
BITGET_SECRET_KEY = os.getenv("BITGET_SECRET_KEY", "")
BITGET_ACCESS_KEY = os.getenv("BITGET_ACCESS_KEY", "")